feat(contextselector): better typing for context selector
