+++ packages/context-selector/src/ContextSelector.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable react/no-multi-comp */
-import { Dropdown, DropdownProps } from '@equinor/fusion-react-searchable-dropdown';
-
-export type ContextSelectorProps = DropdownProps;
+import { Dropdown } from '@equinor/fusion-react-searchable-dropdown';
+import { ContextSelectorProps } from './types';
 
 export const ContextSelector = ({ children, ...props }: React.PropsWithChildren<ContextSelectorProps>): JSX.Element => {
   return <Dropdown {...props}>{children}</Dropdown>;
