+++ storybook/src/stories/context-selector/component.stories.mdx
@@ -1,8 +1,7 @@
 import { Meta, Story, Canvas, ArgsTable, Description } from '@storybook/addon-docs';
-import ChangeLog from '../../components/ChangeLog';
-
 import { ContextProvider, ContextSelector, ContextSelectorProps } from '@equinor/fusion-react-context-selector';
-import { _handleSelect, _exampleResolver } from '../searchable-dropdown/component.helpers.tsx';
+import { _handleSelect, _exampleResolver } from './component.helpers.tsx';
+import ChangeLog from '../../components/ChangeLog';
 
 <Meta title="Data/ContextSelector" />
