+++ storybook/tsconfig.json
@@ -1,10 +1,10 @@
 {
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
-    "rootDir": "src",
+    "rootDir": "./src",
     "resolveJsonModule": true
   },
-  "exclude": ["node_modules"],
+  "exclude": ["./node_modules", "./__tmp__"],
   "references": [
     {
       "path": "../packages/avatar"
